mssql mysql sp 호출 방법이 달라 수정(common-tool 수정 필요), object 타입 캐스팅 오류로 dynamic으로 수정(성능 문제시 강제적으로 하기), 기타 모듈 수정

+++ Service/Service.DB/DBManager.cs
@@ -59,7 +59,7 @@ namespace Service.DB
         public string _dbPW;
         public string _dbIP;
         public string _slaveDBIP;
-        public int _dbPort;
+        public short _dbPort;
         public Byte _threadCount;
         public Dictionary<string, short> _redisDBIndex = new Dictionary<string, short>();
